+++ src/allpairs/allpairs_process.cpp
@@ -81,7 +81,7 @@ void pgr_process_allpairs(
     time_msg(name.c_str(), start_t, clock());
 
     if (!err.str().empty() && (*result_tuples)) {
-        if (*result_tuples) pfree(*result_tuples);
+        pfree(*result_tuples);
         (*result_tuples) = nullptr;
         (*result_count) = 0;
     }