+++ doc/src/release_notes.rst
@@ -45,6 +45,7 @@ To see all issues & pull requests closed by this release see the
 * :issue:`3056`: edwardMoore: use the shortest_path process and driver
 * :issue:`3060`: dagShortestPath: use the shortest_path process and driver
 * :issue:`3064`: Astar: create and use a process and driver for Astar
+* :issue:`3075`: Spanning tree: create and use a process and driver
 
 pgRouting 4.0
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++